chore: splitted functions into different files
